refactor: make replace and replace-all tree shakable
